requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
isort==5.12.0
flake8==6.1.0 